+++ gradle.properties
@@ -1,14 +1,16 @@
 # Done to increase the memory available to gradle.
 org.gradle.jvmargs=-Xmx1G
+
 # Fabric Properties
-# check these on https://modmuss50.me/fabric.html
-minecraft_version=1.17.1
-yarn_mappings=1.17.1+build.61
-loader_version=0.12.0
+	# check these on https://fabricmc.net/versions.html
+	minecraft_version=1.18-pre2
+	yarn_mappings=1.18-pre2+build.1
+	loader_version=0.12.5
+
 # Mod Properties
-mod_version=1.0-SNAPSHOT
-maven_group=firstmod
-archives_base_name=firstmod
+	mod_version = 1.0.0
+	maven_group = com.example
+	archives_base_name = fabric-example-mod
+
 # Dependencies
-# check this on https://modmuss50.me/fabric.html
-fabric_version=0.40.1+1.17
+	fabric_version=0.42.4+1.18
