+++ src/main/java/xyz/mysticgemstones/block/RockTumbler.java
@@ -1,4 +1,4 @@
-package mysticgemstones.mysticgemstones.block;
+package xyz.mysticgemstones.block;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
@@ -20,7 +20,7 @@ public class RockTumbler extends Block {
     public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
 
         if (!world.isClient) {
-            player.sendMessage(new LiteralText("Not yet in game!"), false);
+            player.sendMessage(new LiteralText("Not yet in game!"), true);
         }
         return ActionResult.SUCCESS;
     }
