+++ src/main/java/xyz/mysticgemstones/item/CondensedStarstonePowder.java
@@ -1,4 +1,4 @@
-package mysticgemstones.mysticgemstones.item;
+package xyz.mysticgemstones.item;
 
 import net.fabricmc.fabric.api.object.builder.v1.client.model.FabricModelPredicateProviderRegistry;
 import net.minecraft.entity.Entity;
@@ -7,20 +7,20 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
 import net.minecraft.world.World;
 
-public class StarstonePowder extends Item {
+public class CondensedStarstonePowder extends Item {
 
-    public StarstonePowder(Settings settings) {
+    public CondensedStarstonePowder(Settings settings) {
         super(settings);
     }
 
     public void inventoryTick(ItemStack stack, World world, Entity entity, int slot, boolean selected) {
 
         if (world.getTimeOfDay() > 22350 || world.getTimeOfDay() < 21980) {
-            FabricModelPredicateProviderRegistry.register(MysticGemstonesItem.STARSTONE_POWDER,
+            FabricModelPredicateProviderRegistry.register(MysticGemstonesItem.CONDENSED_STARSTONE_POWDER,
                     new Identifier("shining"), (itemStack, clientWorld, livingEntity, hmmmm) -> 0.0F);
         }
         else if (world.getTimeOfDay() > 21980 && world.getTimeOfDay() < 22350) {
-            FabricModelPredicateProviderRegistry.register(MysticGemstonesItem.STARSTONE_POWDER,
+            FabricModelPredicateProviderRegistry.register(MysticGemstonesItem.CONDENSED_STARSTONE_POWDER,
                     new Identifier("shining"), (itemStack, clientWorld, livingEntity, hmmmm) -> 1.0F);
         }
     }
