+++ src/main/java/xyz/mysticgemstones/item/RawStarstone.java
@@ -1,4 +1,4 @@
-package mysticgemstones.mysticgemstones.item;
+package xyz.mysticgemstones.item;
 
 import net.minecraft.block.Block;
 import net.minecraft.entity.Entity;
