+++ src/main/java/xyz/mysticgemstones/item/charms/AquamarineCharm.java
@@ -1,4 +1,4 @@
-package mysticgemstones.mysticgemstones.item.charms;
+package xyz.mysticgemstones.item.charms;
 
 import net.minecraft.item.Item;
