+++ src/main/resources/fabric.mod.json
@@ -2,27 +2,38 @@
   "schemaVersion": 1,
   "id": "mysticgemstones",
   "version": "${version}",
-  "name": "Aquamatin Mod",
-  "description": "",
-  "authors": [],
-  "contact": {},
-  "license": "MIT",
-  "icon": "assets/firstmod/icon.png",
+
+  "name": "Mystic Gemstones",
+  "description": "Best Mod description ever (Made from fabric example mod for 1.18, https://github.com/FabricMC/fabric-example-mod/tree/1.18)",
+  "authors": [
+    "Domza64",
+    "Jeryux"
+  ],
+  "contact": {
+    "homepage": "https://mysticgemstones.xyz/",
+    "sources": "https://github.com/MysticGemstonesMod/MysticGemstones"
+  },
+
+  "license": "MIT License",
+  "icon": "assets/mysticgemstones/icon.png",
+
   "environment": "*",
   "entrypoints": {
-    "client": [
-      "mysticgemstones.mysticgemstones.client.mysticgemstonesClient"
-    ],
     "main": [
-      "mysticgemstones.mysticgemstones.MysticGemstones"
+      "xyz.mysticgemstones.MysticGemstones"
     ]
   },
   "mixins": [
     "mysticgemstones.mixins.json"
   ],
+
   "depends": {
-    "fabricloader": ">=0.12.0",
+    "fabricloader": ">=0.11.3",
     "fabric": "*",
-    "minecraft": "1.17.1"
+    "minecraft": "1.18.x",
+    "java": ">=17"
+  },
+  "suggests": {
+    "another-mod": "*"
   }
 }
